Word Guessing Mini Game
Project Description: The program randomly selects a word, and the user has to guess it. Initially, all letters of the word are unknown. Additionally, a gallows with a noose is drawn. The user suggests a letter that may be part of the word. If such a letter is in the word, the program reveals the letter as many times as it appears in the word. If the letter is not in the word, a body part of the hangman is added to the gallows, starting with the head. The user continues guessing letters until they guess the entire word. For each incorrect guess, another body part of the hangman is added to the gallows (typically 6 parts: head, torso, 2 arms, and 2 legs).